✨! Use new XDG Base specification for state
Switched the history from using the XDG_CACHE_HOME directory by default
to use XDG_STATE_HOME by default.

This makes sense since cache can (and should be prepared to) be wiped at
any moment and the program functionality should not be hindered by this.

Since we need to retain history through such wipes the newly introduced
state directory is the perfect match for keeping the history file in.

This does constitute a small breaking change for existing histories
which need to be moved to the new directory if they made use of the old
cache directory.

Fixes #5.