🦊 Refactor picker tool choice to use hash map
With the growing amount of picker tools supported out of the box
it seems more practical to have them together in a simple list
that can grow and through which we can always iterate instead
of a cumbersome growing if-else chain.

This commit moves the pick tool selection to be stored in
a bash hash map instead.