📖 Bump license copyright year
Bumping the year again. I am honestly not entirely sure what the best or
correct way to display this is - so I am simply copying what the `curl`
project is doing and stating the whole period of me working on it.

Perhaps one day (should we get more contributors or much more popular)
it may also make sense to copy its `THANKS` file to keep a stable copy
of my contributing helpers.