🧪 Adapt tests to custom type cmd fixes
The fixes in 65efc27 require some adaption of the test code since we now
actually, correctly, pass through any chosen emoji. Thus, the emoji will
also be 'typed' (or in the case of the tests simply echo'd) out on the
command line which we now test for with this commit.