🐛 Fix removed unicode download url for emoji
This is a quick fix for the missing 'emoji-test.txt' file in the Public
directory of the unicode.org website for the latest emoji list.

We fix it by pinning the download location to unicode version 16.0,
which still had the required file in its correct location.

This should be a temporary fix until the unicode website either adds the
correct file to the updated latest version, or we switch our downloading
logic.