🧪 Fix ci test suite
Instead of using the Github action "bat-action", we simply instantiate
the repository with submodules included. Since we have the bats-suite
included as various submodules, this gives us everything we need. Now we
can simlpy run the locally given bats binary and run it on the test
directory. All library dependencies are loaded from the correct dirs in
the test setups themselves.